Ensure both footnotes and reviews split correctly and renumbered under sectionsplit: https://github.com/metanorma/iso-10303/issues/561
