+++ Gemfile.devel
@@ -1,6 +1,5 @@
-gem "isodoc", git: "https://github.com/metanorma/isodoc", branch: "features/anchor"
-gem "relaton-iev", git: "https://github.com/relaton/relaton-iev", branch: "features/anchor"
-gem "metanorma-standoc", git: "https://github.com/metanorma/metanorma-standoc", branch: "features/anchor"
-gem "metanorma-iso", git: "https://github.com/metanorma/metanorma-iso", branch: "features/anchor"
+gem "isodoc", git: "https://github.com/metanorma/isodoc", branch: "main"
+gem "metanorma-standoc", git: "https://github.com/metanorma/metanorma-standoc", branch: "main"
+gem "metanorma-iso", git: "https://github.com/metanorma/metanorma-iso", branch: "main"
 gem "metanorma-utils", git: "https://github.com/metanorma/metanorma-utils", branch: "main"
-gem "metanorma-cli"
+#gem "metanorma-cli"
