+++ spec/fixtures/collection/footnotes.yml
@@ -1,0 +1,48 @@
+# directives are YAML-specific instructions
+directives:
+  - documents-inline
+  # to inject the XML into the collection manifest; 
+  # - documents-external 
+  # keeps them outside
+bibdata:
+  title:
+    type: title-main
+    language: en
+    content: ISO Collection 1
+  type: collection
+  docid:
+    type: iso
+    id: ISO 12345
+  edition: 
+    content: 1
+  date:
+    - type: created
+      value: "2020"
+    - type: issued
+      value: "2020"
+  copyright:
+    owner:
+      name: International Organization for Standardization
+      abbreviation: ISO
+    from: "2020"
+manifest:
+  level: collection
+  title: ISO Collection 
+  manifest:
+    - level: subcollection
+      title: Standards
+      docref:
+        - fileref: footnotes.xml
+          identifier: ISO 17301-1:2016
+          sectionsplit: true
+prefatory-content: 
+|
+
+ == Clause
+ Welcome to our collection
+
+final-content:
+|
+
+ == Exordium
+ Hic explicit
