+++ spec/spec_helper.rb
@@ -13,6 +13,7 @@ require "rspec-command"
 require "mnconvert"
 require "mn2pdf"
 require "xml-c14n"
+require_relative "support/uuid_mock"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -116,8 +117,35 @@ HDR
 
 def strip_guid(xml)
   xml.gsub(%r{ id="_[^"]+"}, ' id="_"')
-    .gsub(%r{ semx-id="[^"]*"}, '')
+    .gsub(%r{ semx-id="[^"]*"}, "")
     .gsub(%r{ target="_[^"]+"}, ' target="_"')
+end
+
+def cleanup_guid(content)
+  content
+    .gsub(%r{cid:#{GUID}}o, "cid:_")
+    .gsub(%r{ semx-id="[^"]*"}o, "")
+    .gsub(%r{ id="_#{GUID}"}o, ' id="_"')
+    .gsub(%r{ target="_#{GUID}"}o, ' name="_"')
+    .gsub(%r{ source="_#{GUID}"}o, ' source="_"')
+    .gsub(%r{ original-id="_#{GUID}"}o, ' original-id="_"')
+    .gsub(%r{ name="_#{GUID}"}o, ' name="_"')
+    .gsub(%r{_Toc[0-9]{9}}o, "_Toc")
+end
+
+def read_and_cleanup(file)
+  content = File.read(file, encoding: "UTF-8").gsub(
+    /(?<=<fetched>)\d{4}-\d{2}-\d{2}/, Date.today.to_s
+  )
+  cleanup_id content
+end
+
+def cleanup_id(content)
+  content.gsub(/(?<=<p id=")[^"]+/, "")
+    .gsub(%r{data:image/svg\+xml[^<"']+}, "data:image/svg+xml")
+    .gsub(%r{data:image/png[^<"']+}, "data:image/png")
+    .gsub(/ schema-version="[^"]+"/, "")
+    .gsub(%r{<identifier>#{GUID}</identifier>}o, "<identifier>_</identifier>")
 end
 
 def mock_pdf
